+++ rules/variables.js
@@ -1,8 +1,7 @@
 var rules = {
-    'no-undef': 'warn',
-    'no-undefined': 'error',
-    'no-unused-vars': 'warn',
-    'no-use-before-define': 'warn',
+  'no-undef': 'warn',
+  'no-undefined': 'warn',
+  'no-unused-vars': 'warn',
 };
 
 module.exports = rules;