+++ index.js
@@ -1,10 +1,11 @@
 var rules = require('./rules');
 
 module.exports = {
-    env: {
-        browser: true,
-        node: true,
-        es6: true,
-    },
-    rules: rules,
+  env: {
+    browser: true,
+    node: true,
+    es6: true,
+  },
+  rules: rules,
+  plugins: ['react'],
 };