+++ rules/react/jsx.js
@@ -1,0 +1,13 @@
+var rules = {
+  'react/jsx-filename-extension': 'error',
+  'react/jsx-key': 'error',
+  'react/jsx-no-bind': 'error',
+  'react/jsx-no-comment-textnodes': 'warn',
+  'react/jsx-no-duplicate-props': 'warn',
+  'react/jsx-no-target-blank': 'error',
+  'react/jsx-no-undef': 'error',
+  'react/jsx-uses-react': 'error',
+  'react/jsx-uses-vars': 'error',
+};
+
+module.exports = rules;